backend/requirements.txt
fastapi>=0.111.0
requests>=2.31.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
uvicorn>=0.30.1
pydantic>=2.7.4
